LayoutTests/js/dom/enter-dictionary-indexing-mode-with-blank-indexing-type-expected.txt

Tests that putting an object into dictionary mode when it has a blank indexing type doesn't cause us to crash.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Object.isFrozen(Array.prototype) is true
PASS successfullyParsed is true

TEST COMPLETE
